# docs/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/users": {
      "get": {
        "operationId": "UsersController_findAll",
        "summary": "Liste des utilisateurs",
        "parameters": [
          {
            "name": "role",
            "required": false,
            "in": "query",
            "schema": {
              "enum": [
                "ADMIN",
                "SUPER_MANAGER",
                "MANAGER",
                "DRIVER",
                "OWNER"
              ],
              "type": "string"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "users"
        ],
        "security": [
          {
            "JWT": []
          }
        ]
      },
      "post": {
        "operationId": "UsersController_create",
        "summary": "Créer un utilisateur",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "users"
        ],
        "security": [
          {
            "JWT": []
          }
        ]
      }
    },
    "/api/v1/users/{id}": {
      "get": {
        "operationId": "UsersController_findById",
        "summary": "Détail utilisateur",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "users"
        ],
        "security": [
          {
            "JWT": []
          }
        ]
      },
      "patch": {
        "operationId": "UsersController_update",
        "summary": "Modifier un utilisateur",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "users"
        ],
        "security": [
          {
            "JWT": []
          }
        ]
      },
      "delete": {
        "operationId": "UsersController_remove",
        "summary": "Supprimer (soft delete) un utilisateur",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "users"
        ],
        "security": [
          {
            "JWT": []
          }
        ]
      }
    },
    "/api/v1/users/{id}/password": {
      "patch": {
        "operationId": "UsersController_changePassword",
        "summary": "Changer son mot de passe",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangePasswordDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "users"
        ],
        "security": [
          {
            "JWT": []
          }
        ]
      }
    },
    "/api/v1/users/{id}/suspend": {
      "patch": {
        "operationId": "UsersController_suspend",
        "summary": "Suspendre un utilisateur",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "users"
        ],
        "security": [
          {
            "JWT": []
          }
        ]
      }
    },
    "/api/v1/users/{id}/activate": {
      "patch": {
        "operationId": "UsersController_activate",
        "summary": "Réactiver un utilisateur",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "users"
        ],
        "security": [
          {
            "JWT": []
          }
        ]
      }
    },
...
    "/api/v1/media/upload": {
      "post": {
        "operationId": "MediaController_upload",
        "summary": "Upload un fichier — crée un MediaAsset",
        "parameters": [],
        "requestBody": {
          "required": true,
          "description": "Fichier multipart + métadonnées JSON",
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "string"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "media"
        ],
        "security": [
          {
            "JWT": []
          }
        ]
      }
    },
    "/api/v1/media/photos": {
      "post": {
        "operationId": "MediaController_createPhoto",
        "summary": "Créer un wrapper Photo depuis un MediaAsset",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePhotoDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "media"
        ],
        "security": [
          {
            "JWT": []
          }
        ]
      }
    },
    "/api/v1/media/photo-missions": {
      "post": {
        "operationId": "MediaController_createPhotoMission",
        "summary": "Créer une mission photo pour un chauffeur",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePhotoMissionDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "media"
        ],
        "security": [
          {
            "JWT": []
          }
        ]
      },
      "get": {
        "operationId": "MediaController_findPhotoMissions",
        "summary": "Lister les missions photo",
        "parameters": [
          {
            "name": "vehicleId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "driverId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "media"
        ],
        "security": [
          {
            "JWT": []
          }
        ]
      }
    },
    "/api/v1/media/photo-missions/{id}/submit": {
      "post": {
        "operationId": "MediaController_submitPhotoMission",
        "summary": "Soumettre une mission photo (chauffeur)",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SubmitPhotoMissionDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "media"
        ],
        "security": [
          {
            "JWT": []
          }
        ]
      }
    },
    "/api/v1/media/photo-missions/{id}/validate": {
      "post": {
        "operationId": "MediaController_validatePhotoMission",
        "summary": "Valider une mission photo soumise",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "media"
        ],
        "security": [
          {
            "JWT": []
          }
        ]
      }
    },
    "/api/v1/media/photo-missions/{id}": {
      "get": {
        "operationId": "MediaController_findPhotoMissionById",
        "summary": "Détail d'une mission photo",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "media"
        ],
        "security": [
          {
            "JWT": []
          }
        ]
      }
    },
    "/api/v1/media/{id}/url": {
      "get": {
        "operationId": "MediaController_getSignedUrl",
        "summary": "Obtenir l'URL (signée) d'un MediaAsset",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "media"
        ],
        "security": [
          {
            "JWT": []
          }
        ]
      }
    },
    "/api/v1/media/{id}": {
      "get": {
        "operationId": "MediaController_findById",
        "summary": "Détail d'un MediaAsset",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "media"
        ],
        "security": [
          {
            "JWT": []
          }
        ]
      }
    },
    "/api/v1/media": {
      "get": {
        "operationId": "MediaController_findAll",
        "summary": "Lister les MediaAssets (filtrés par entité)",
        "parameters": [
          {
            "name": "entityType",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "entityId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "mediaType",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "media"
        ],
        "security": [
          {
            "JWT": []
          }
        ]
      }
    },
...
  },
  "info": {
    "title": "Fleet Platform API",
    "description": "API de gestion de flotte VTC — Fleet Platform V2",
    "version": "2.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "JWT": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CreateUserDto": {
        "type": "object",
        "properties": {
          "firstName": {
            "type": "string",
            "example": "Mamadou"
          },
          "lastName": {
            "type": "string",
            "example": "Diallo"
          },
          "email": {
            "type": "string",
            "example": "mamadou.diallo@fleet.local"
          },
          "phone": {
            "type": "string",
            "example": "+221701234567"
          },
          "password": {
            "type": "string",
            "minLength": 8,
            "example": "MonMotDePasse123!"
          },
          "role": {
            "type": "string",
            "enum": [
              "ADMIN",
              "SUPER_MANAGER",
              "MANAGER",
              "DRIVER",
              "OWNER"
            ]
          }
        },
        "required": [
          "firstName",
          "lastName",
          "password",
          "role"
        ]
      },
      "UpdateUserDto": {
        "type": "object",
        "properties": {
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "phone": {
            "type": "string"
          },
          "avatarUrl": {
            "type": "string"
          },
          "fcmToken": {
            "type": "string"
          }
        }
      },
      "ChangePasswordDto": {
        "type": "object",
        "properties": {
          "currentPassword": {
            "type": "string"
          },
          "newPassword": {
            "type": "string",
            "minLength": 8
          }
        },
        "required": [
          "currentPassword",
          "newPassword"
        ]
      },
...
      "CreatePhotoDto": {
        "type": "object",
        "properties": {
          "mediaAssetId": {
            "type": "string",
            "description": "ID du MediaAsset déjà uploadé"
          },
          "missionId": {
            "type": "string",
            "description": "ID de la mission photo liée"
          },
          "inspectionId": {
            "type": "string",
            "description": "ID de l'inspection liée"
          },
          "captureContext": {
            "type": "string",
            "description": "Contexte de capture (DAMAGE, ODOMETER, SCENE…)"
          }
        },
        "required": [
          "mediaAssetId"
        ]
      },
      "CreatePhotoMissionDto": {
        "type": "object",
        "properties": {
          "vehicleId": {
            "type": "string",
            "description": "ID du véhicule concerné"
          },
          "driverId": {
            "type": "string",
            "description": "ID du chauffeur assigné"
          },
          "type": {
            "type": "string",
            "enum": [
              "REGULAR_VEHICLE",
              "DASHBOARD_MILEAGE",
              "BREAKDOWN",
              "ACCIDENT",
              "INSPECTION",
              "TOWING"
            ]
          },
          "dueDate": {
            "type": "string",
            "description": "Date limite de soumission (ISO 8601)"
          },
          "incidentId": {
            "type": "string",
            "description": "ID de l'incident déclencheur"
          }
        },
        "required": [
          "vehicleId",
          "driverId",
          "type",
          "dueDate"
        ]
      },
      "SubmitPhotoMissionDto": {
        "type": "object",
        "properties": {
          "photoIds": {
            "description": "Liste des IDs de Photo à lier à cette mission",
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "photoIds"
        ]
      },
...
    }
  }
}